Namespace all the modules under HTML::Mixin.
